feat: splits vue widget into components
