+++ vite/src/components/HiddenInput.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+const props = defineProps<{
+  uploadedImageLocation: string | null;
+}>();
+</script>
+
+<template>
+  <input type="hidden" name="Project[image_location]" :value="props.uploadedImageLocation" />
+</template>
