+++ vite/src/config.ts
@@ -1,0 +1,9 @@
+const maxHeight = 60;
+const maxSize = 1000000;
+const maxSizeMb = maxSize / 1e6;
+
+export const config = {
+  maxHeight,
+  maxSize,
+  maxSizeMb
+}
